move checkpoint views for submodule datatypes to submodules
