+++ ixmp4/core/checkpoint.py
@@ -13,6 +13,8 @@ from ixmp4.data.checkpoint.exceptions import (
 from ixmp4.data.checkpoint.service import CheckpointService
 
 from .base import BaseBackendFacade, BaseServiceFacade
+from .iamc.checkpoint import CheckpointIamcData
+from .optimization.checkpoint import CheckpointOptimizationData
 
 if TYPE_CHECKING:
     from .run import Run
@@ -21,226 +23,6 @@ _VERSIONING_NOT_SUPPORTED_MSG = (
     "Checkpoint data views require PostgreSQL versioning support. "
     "This feature is not available on the current backend."
 )
-
-
-class CheckpointScalarView(BaseBackendFacade):
-    """Read-only view of optimization scalars at a checkpoint."""
-
-    def __init__(self, backend: Backend, run: "Run", checkpoint: Checkpoint) -> None:
-        super().__init__(backend)
-        self._run = run
-        self._checkpoint = checkpoint
-
-    def tabulate(self, **kwargs: Any) -> pd.DataFrame:
-        """Tabulate scalar versions at this checkpoint.
-
-        Returns
-        -------
-        :class:`pandas.DataFrame`
-            DataFrame with the scalar version columns.
-        """
-        if self._checkpoint.transaction__id is None:
-            raise OperationNotSupported(_VERSIONING_NOT_SUPPORTED_MSG)
-        return self._backend.optimization.scalars.tabulate_versions(
-            run__id=self._run.id,
-            valid_at_transaction=self._checkpoint.transaction__id,
-            **kwargs,
-        )
-
-
-class CheckpointTableView(BaseBackendFacade):
-    """Read-only view of optimization tables at a checkpoint."""
-
-    def __init__(self, backend: Backend, run: "Run", checkpoint: Checkpoint) -> None:
-        super().__init__(backend)
-        self._run = run
-        self._checkpoint = checkpoint
-
-    def tabulate(self, **kwargs: Any) -> pd.DataFrame:
-        """Tabulate table versions at this checkpoint.
-
-        Returns
-        -------
-        :class:`pandas.DataFrame`
-            DataFrame with the table version columns.
-        """
-        if self._checkpoint.transaction__id is None:
-            raise OperationNotSupported(_VERSIONING_NOT_SUPPORTED_MSG)
-        return self._backend.optimization.tables.tabulate_versions(
-            run__id=self._run.id,
-            valid_at_transaction=self._checkpoint.transaction__id,
-            **kwargs,
-        )
-
-
-class CheckpointParameterView(BaseBackendFacade):
-    """Read-only view of optimization parameters at a checkpoint."""
-
-    def __init__(self, backend: Backend, run: "Run", checkpoint: Checkpoint) -> None:
-        super().__init__(backend)
-        self._run = run
-        self._checkpoint = checkpoint
-
-    def tabulate(self, **kwargs: Any) -> pd.DataFrame:
-        """Tabulate parameter versions at this checkpoint.
-
-        Returns
-        -------
-        :class:`pandas.DataFrame`
-            DataFrame with the parameter version columns.
-        """
-        if self._checkpoint.transaction__id is None:
-            raise OperationNotSupported(_VERSIONING_NOT_SUPPORTED_MSG)
-        return self._backend.optimization.parameters.tabulate_versions(
-            run__id=self._run.id,
-            valid_at_transaction=self._checkpoint.transaction__id,
-            **kwargs,
-        )
-
-
-class CheckpointEquationView(BaseBackendFacade):
-    """Read-only view of optimization equations at a checkpoint."""
-
-    def __init__(self, backend: Backend, run: "Run", checkpoint: Checkpoint) -> None:
-        super().__init__(backend)
-        self._run = run
-        self._checkpoint = checkpoint
-
-    def tabulate(self, **kwargs: Any) -> pd.DataFrame:
-        """Tabulate equation versions at this checkpoint.
-
-        Returns
-        -------
-        :class:`pandas.DataFrame`
-            DataFrame with the equation version columns.
-        """
-        if self._checkpoint.transaction__id is None:
-            raise OperationNotSupported(_VERSIONING_NOT_SUPPORTED_MSG)
-        return self._backend.optimization.equations.tabulate_versions(
-            run__id=self._run.id,
-            valid_at_transaction=self._checkpoint.transaction__id,
-            **kwargs,
-        )
-
-
-class CheckpointVariableView(BaseBackendFacade):
-    """Read-only view of optimization variables at a checkpoint."""
-
-    def __init__(self, backend: Backend, run: "Run", checkpoint: Checkpoint) -> None:
-        super().__init__(backend)
-        self._run = run
-        self._checkpoint = checkpoint
-
-    def tabulate(self, **kwargs: Any) -> pd.DataFrame:
-        """Tabulate variable versions at this checkpoint.
-
-        Returns
-        -------
-        :class:`pandas.DataFrame`
-            DataFrame with the variable version columns.
-        """
-        if self._checkpoint.transaction__id is None:
-            raise OperationNotSupported(_VERSIONING_NOT_SUPPORTED_MSG)
-        return self._backend.optimization.variables.tabulate_versions(
-            run__id=self._run.id,
-            valid_at_transaction=self._checkpoint.transaction__id,
-            **kwargs,
-        )
-
-
-class CheckpointIndexSetView(BaseBackendFacade):
-    """Read-only view of optimization indexsets at a checkpoint."""
-
-    def __init__(self, backend: Backend, run: "Run", checkpoint: Checkpoint) -> None:
-        super().__init__(backend)
-        self._run = run
-        self._checkpoint = checkpoint
-
-    def tabulate(self, **kwargs: Any) -> pd.DataFrame:
-        """Tabulate indexset versions at this checkpoint.
-
-        Returns
-        -------
-        :class:`pandas.DataFrame`
-            DataFrame with the indexset version columns.
-        """
-        if self._checkpoint.transaction__id is None:
-            raise OperationNotSupported(_VERSIONING_NOT_SUPPORTED_MSG)
-        return self._backend.optimization.indexsets.tabulate_versions(
-            run__id=self._run.id,
-            valid_at_transaction=self._checkpoint.transaction__id,
-            **kwargs,
-        )
-
-
-class CheckpointOptimizationData(BaseBackendFacade):
-    """Read-only view of all optimization data for a run at a checkpoint."""
-
-    scalars: CheckpointScalarView
-    tables: CheckpointTableView
-    parameters: CheckpointParameterView
-    equations: CheckpointEquationView
-    variables: CheckpointVariableView
-    indexsets: CheckpointIndexSetView
-
-    def __init__(self, backend: Backend, run: "Run", checkpoint: Checkpoint) -> None:
-        super().__init__(backend)
-        self.scalars = CheckpointScalarView(backend, run, checkpoint)
-        self.tables = CheckpointTableView(backend, run, checkpoint)
-        self.parameters = CheckpointParameterView(backend, run, checkpoint)
-        self.equations = CheckpointEquationView(backend, run, checkpoint)
-        self.variables = CheckpointVariableView(backend, run, checkpoint)
-        self.indexsets = CheckpointIndexSetView(backend, run, checkpoint)
-
-
-class CheckpointIamcData(BaseBackendFacade):
-    """Read-only view of IAMC data for a run at a checkpoint."""
-
-    def __init__(self, backend: Backend, run: "Run", checkpoint: Checkpoint) -> None:
-        super().__init__(backend)
-        self._run = run
-        self._checkpoint = checkpoint
-
-    def tabulate(self, **kwargs: Any) -> pd.DataFrame:
-        """Tabulate IAMC data at this checkpoint, in standard IAMC format.
-
-        Returns a DataFrame with columns: region, variable, unit, year
-        (or time), and value. Data reflects the state of the run at the
-        checkpoint's transaction.
-
-        Returns
-        -------
-        :class:`pandas.DataFrame`
-            IAMC-formatted data at checkpoint time.
-
-        Raises
-        ------
-        :class:`ixmp4.base_exceptions.OperationNotSupported`
-            If versioning is not supported on this backend (e.g. SQLite).
-        """
-        from ixmp4.core.iamc.data import _convert_to_std_format
-
-        if self._checkpoint.transaction__id is None:
-            raise OperationNotSupported(_VERSIONING_NOT_SUPPORTED_MSG)
-
-        dp_df = self._backend.iamc.datapoints.tabulate_versions(
-            timeseries={"run__id": self._run.id},
-            valid_at_transaction=self._checkpoint.transaction__id,
-        )
-
-        if dp_df.empty:
-            return pd.DataFrame(columns=["region", "variable", "unit", "year", "value"])
-
-        # Enrich version datapoints with region/variable/unit from live timeseries
-        ts_df = self._backend.iamc.timeseries.tabulate(
-            run__id=self._run.id,
-            run={"default_only": False},
-        )
-        ts_df = ts_df.rename(columns={"id": "time_series__id"})[
-            ["time_series__id", "region", "variable", "unit"]
-        ]
-        merged = dp_df.merge(ts_df, on="time_series__id", how="inner")
-        return _convert_to_std_format(merged, join_runs=False, join_run_id=False)
 
 
 class CheckpointView(BaseBackendFacade):
